support emacs 29.1
In emacs 29.1 format-time-string no longer wraps time strings
in their corresponding square brackets ("[", "]"). This breaks
org-clock-split in its current form.
This adds a function that provides a proper inactive time string, with the
wrapped square brackets if it does not yet have them (meaning it'll support both
emacs 28 and 29+).

Also removes one superfluous test that is already handled in the larger
duplicated org-clock-split-split-line-test.