correct test files
In emacs-29.1 timestamps with incorrect day names will be automatically updated
as part of the timestamp formatting functions, this revealed the test cases
where the selected day was incorrect (2019-12-17 was a Tuesday not a Wednesday).
This updates both the expected values and the source values of the test
cases (source values are updated because the tests would otherwise fail on emacs-28.2).